Bumping SDK, driver, and SM for user-space ELF loading and TOR support
Now the eapp and the runtime ELFs are parsed by the user space ELF
library, which provides a way more flexible enclave initialization.
The driver is no longer responsible for ELF loading, but it will still
be in charge of managing the physical memory at initialization.
The user-level host parses ELF, and calls proper ioctls to initialize the
enclave prior to finalizing the enclave measurement.

In addition, the SM now handles unaligned CMA regions by looking at the
address and the size of the enclave memory allocated by the OS.
To do this, the SM uses two pmp entries encoded in TOR, which is
suboptimal. But this works well for a small number of concurrent
enclaves.

This resolves #55, and #54.