Add ALLOC_VSPACE ioctl for linearized mapping
A new ioctl is added for user to allocate page tables prior to loading
the ELF, so that the ELF is linearly mapped to the physical memory.
Linear mapping is not mandatory, but is strongly recommended for runtimes
that manage their own page tables.
For example, SeL4 calculates the physical address of its own pages based
on the offset of the virtual address with respect to the physical address,
so the host should call this ioctl to make the offset consistent over
all pages.