Change shared buffer VA to avoid using the same PTE with the runtime
We're in the middle of transition, which makes runtime map the
untrusted buffer by itself.
Since the current address 0xffffffffe0000000 uses the same top-level PTE
for the runtime itself, copy_page_table will not correctly copy the PTE
created by the OS to the new page table.
Temporarily, we just change the address to avoid this.

This line will completely removed when the transition ends