[all] Massive restructuring of SDK (#55)
(1) This changes the directory structure to be more sane.
- Use CMake for all examples
  Remove Makefile and vault.sh
  Remove rts and let CMake handle the runtime with ExternalProject_add
- Directory lib/app/include --> include/app
- Directory lib/host/include --> include/host
- Directory lib/edge/include --> include/edge
- Directory lib/verifier/include --> include/verifier
- Directory common/include --> include/common
- Directory lib/app/src --> src/app
- Directory lib/host/src --> src/host
- Directory lib/edge/src --> src/edge
- Directory lib/verifier/src --> src/verifier
- Directory common/src --> src/common

(2) This change completely separates the source and the build files.
Also, it makes it wildly easier to "install" the headers and the
libraries to a designated directory configured with a CMake flag,
OUTPUT_DIR. Now, KEYSTONE_SDK_DIR should point to this "install" directory instead
of the source directory. Thus, it may require runtime and keystone repo updates.

(3) Auto-format everything with clang-format and cpplint. `make format` to run them

(4) Minor changes to cpp sources/headers

(5) Include auto formating to travis CI