Remove the dependency from image-deps to tests.
image-deps is not really a target specifically for testing, but tests is a target for tests. Generally, it is a good practice to not let non-test targets depend on test target, and only let test-only targets depend on test targets.

For this specific situation, we had a dependency because we wanted to build the test binaries into the image, but this is not really a good idea, as production images will also contain test binaries. Instead, we will be scp'ing the test binaries into the QEMU VM when run-tests is invoked.