Writes the remote attestation tutorial. (#250)
* Writes the remote attestation tutorial.

This is dependent on the pull request on keystone-sdk: https://github.com/keystone-enclave/keystone-sdk/pull/77. DO NOT merge before merging https://github.com/keystone-enclave/keystone-sdk/pull/77.

Also adds the attestation example to the continous integration tests to prevent regression.

* Make the main repo track the recently updated keystone-sdk master.

* Copy fw_payload.bin to /root when making the overlay_root for tests.

This is so that the attestation example can use this fw_payload.bin
for hash calculation.

* Remove the dependency from buildroot to image-deps.

* Copy fw_payload.bin only if initramfs is false.

See keystone-enclave/keystone#251 for the rationale behind this.