Specify input format for .options_log (#65)
For some reason, the wrapping SDK build system does not know about the
input format for the .options_log file. Therefore, we make this explicit
by including a flag that specifies the options_log as raw binary. This
has the desired effect of copying the configuration options into the
final executable.

I also propose bumping the Eyrie release to v1.2.1 after this