Fix off-by-one in the driver (#318)
The existing driver code checks whether an enclave's requested size is
less than or equal to some `MAX_ORDER` defined in the Linux kernel, and
chooses the kernel allocator to use based on this. However, this should
be a strict less than check instead. If we attempt to instantiate an
enclave of size exactly `MAX_ORDER`, we will trigger a bug condition in
Linux's `page_alloc.c` in the `__alloc_pages` function on line 5406.
This condition is enforced when the requested order is greater than or
equal to `MAX_ORDER`. This PR changes the driver such that we do not
attempt to use the wrong allocator and no longer see these bug messages.