Use base location of runtime source files, not compiled location (#319)
When debugging in GDB, it can be helpful to use tools such as `layout
src` in order to use the source files for a specific runtime. However,
the embedded debugging information points to the copied source directory
in the build directory rather than the canonical source under
`runtime/`. This PR adds a compile flag, `-fdebug-prefix-map`, which
embeds debugging information that points to the canonical runtime
source. This is useful in exotic (but seen!) debugging configurations
where the runtime build directory may have moved or no longer exists.