Fix CMake error when top-level directory not `keystone` (#336)
This PR resolves the issue where the `runtime` directory cannot be found
when the top-level directory is not named `keystone`. We resolve the
issue by relying on CMake project names instead of directory names. This
change was verified by successfully building the project in a renamed
top-level directory. For more details, see #335.

Resolves #335.