Build toolchain in Buildroot, rather than downloading a prebuilt one
This is a fairly major change, which completely removes the prebuilt toolchains
from the Keystone build. By doing so, we start the process of centralizing the
build on Buildroot. For most of the subprojects, this change is relatively
straightforward. We create a new CMake target named "buildroot_toolchain" which
only builds the compiler. Most targets then depend on this one, and PATHs are
passed appropriately to ensure that the compiler can be found. The only wrinkle
in this setup is the ExternalProjects for the SDK and examples. CMake does not
make it possible to pass environment variables to such projects, so both the
PATH information as well as the SDK location must be passed as CMake arguments.
This flips some of the $ENV{KEYSTONE_SDK_DIR} / ${KEYSTONE_SDK_DIR} logic in
these builds. This commit therefore also removes the source.sh requirement,
since all relevant information is passed through CMake command flags now.

Additionally, we somewhat modify the runtime since eapps which use the glibc
provided by the toolchain expert certain behavior. Specifically, we bump the
kernel version reported by the runtime's uname syscall and also implement the
mprotect syscall, which is now required by the libc.