Add jq as a dependency in docs (#391)
This little utility [is used for
`linux-configure`](https://github.com/keystone-enclave/keystone/blob/master/Makefile#L107),
and wasn't included as a part of my desktop version of Ubuntu.

Maybe this should be included in your Dockerfile? I'm not using the
image so someone should check if it's there or not.