In-enclave ELF loading (#326)
Changes are in SDK, driver, SM, and runtime. Loader is added to runtime.

A contributor-private branch based on this fully passes `tests` on CVA6.

### Changes & Rationale
Previously, the host SDK was required to produce the same mapping as the
developer SDK. This complicates measurement, keeping SDK versions in
sync, alternate loading features a developer may want, and the Keystone
code in general.

- Enclaves now need "trusted loader" that loads the runtime and enclave
app binaries
- ELF parsing/loading was removed from SDK and added to `runtime/loader`
and `runtime/loader-binary`
- Enclave measurement is now just the measurement of the binaries
concatenated

The new design creates a much simpler code design where the host simply
passes the required elf files to the enclave memory and those files are
measured. The loading code lives in the enclave that makes it simpler
and more standard. Trust assumptions are unchanged because the developer
needs to have correct loading code either way. The developer can now
also have completely custom loading logic that is independent from what
the host supports.

### Outstanding issues
1. Improving enclave measurement: untrusted and free memory are not
measured in, so host can deny resources with measurement not reflecting
that.
2. Minor memory leak: when elf file pages are not guaranteed to be
non-colliding, a new page is allocated instead of that page being used
to one of the colliding program headers.
3. (Unrelated) Buildroot system bug: `fw_bin` is not properly set in the
examples build, added temporary fix.

### Refactor needs
1. `linux-driver` structures
2. No longer needed in SDK: simulated, Memory
3. Make structures consistent across: `sm - open_sbi - driver - ioctl -
host`

---------

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>
Co-authored-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>