Implement new CI on Github Actions (#399)
This PR creates a new efficient CI system for the Buildroot-based build
system introduced a while ago. It includes full build tests for each
supported platform (currently `generic` and `unmatched`, with `cva6` and
`mpfs` landing soon) as well as full-stack runtime tests for the 32 and
64-bit generic platforms. CI times are kept low by heavily relying on
Github Actions caching, where we save the latest compiler cache and
Buildroot package cache at the end of each run. On average, the
`generic` platforms build in 10 minutes while the `unmatched` platform
builds in 13 (with a hot cache). With a cold cache, the builds take
about an hour but will then refresh the caches so that future ones take
much less time.